Make acceptation of new drawings easier

+++ ActionTree/tests/Report.py
@@ -37,7 +37,7 @@ class Report(unittest.TestCase):
 
     # Expect several digests because cairo may not produce exactly the same file on
     # all platorms and versions
-    def __checkDrawing(self, a, width, expectedDigests):
+    def __checkDrawing(self, a, width):
         testName = None
         for (_, _, functionName, _) in traceback.extract_stack():
             if functionName.startswith("test"):
@@ -58,10 +58,10 @@ class Report(unittest.TestCase):
         image.write_to_png(f)
         digest = hashlib.md5(f.getvalue()).hexdigest()
         fileName = os.path.join("ActionTree", "tests", "drawings", testName + "." + digest + ".png")
-        with open(fileName, "wb") as png:
-            png.write(f.getvalue())
-        if digest not in expectedDigests.values():
-            self.assertTrue(False, "Check file " + fileName + ". If it is OK, modify test " + testName + " to accept digest " + digest)
+        if not os.path.exists(fileName):
+            with open(fileName, "wb") as png:
+                png.write(f.getvalue())
+            self.assertTrue(False, "Check file " + fileName + ".")
         f.close()
 
     def __createMockedAction(self, name, label, dependencies, beginTime, endTime, status):
@@ -76,35 +76,35 @@ class Report(unittest.TestCase):
     def testComplexLabel(self):
         a = self.__createMockedAction("a", ("a", "complex", [42, "label"]), [], 10.5, 13.5, ActionTree.Action.Successful)
 
-        self.__checkDrawing(a, 400, {"Python 2.7, Cygwin": "5f3cfe117ab21440c9d7bc0f8d1c9e70"})
+        self.__checkDrawing(a, 400)
 
     def testOneSuccessfulAction(self):
         a = self.__createMockedAction("a", "label", [], 10.5, 13.5, ActionTree.Action.Successful)
 
-        self.__checkDrawing(a, 200, {"Python 2.7, Cygwin": "80da16a2a8c0b5a3cfe71566c1d1a750"})
+        self.__checkDrawing(a, 200)
 
     def testOneFailedAction(self):
         a = self.__createMockedAction("a", "label", [], 10.5, 13.5, ActionTree.Action.Failed)
 
-        self.__checkDrawing(a, 200, {"Python 2.7, Cygwin": "62ad8ef401b9bc37104e58a597e04d13"})
+        self.__checkDrawing(a, 200)
 
     def testOneCanceledAction(self):
         a = self.__createMockedAction("a", "label", [], 10.5, 13.5, ActionTree.Action.Canceled)
 
-        self.__checkDrawing(a, 200, {"Python 2.7, Cygwin": "081c688ac8678ec4ea5f1621c4e8044a"})
+        self.__checkDrawing(a, 200)
 
     def testTwoChainedActions(self):
         a1 = self.__createMockedAction("a1", "a1", [], 10.5, 13.5, ActionTree.Action.Successful)
         a2 = self.__createMockedAction("a2", "a2", [a1], 14.0, 15.5, ActionTree.Action.Successful)
 
-        self.__checkDrawing(a2, 200, {"Python 2.7, Cygwin": "7ed1b75b6c45b594e643bf0128158d83"})
+        self.__checkDrawing(a2, 200)
 
     def testActionWithTwoDependencies(self):
         a1 = self.__createMockedAction("a1", "a1", [], 10.5, 13.5, ActionTree.Action.Successful)
         a2 = self.__createMockedAction("a2", "a2", [], 11.5, 14.0, ActionTree.Action.Successful)
         a3 = self.__createMockedAction("a3", "a3", [a1, a2], 14.0, 15.5, ActionTree.Action.Successful)
 
-        self.__checkDrawing(a3, 200, {"Python 2.7, Cygwin - 1": "e0ee94d2b3e57b05f13d8b0f459c0a2c", "Python 2.7, Cygwin - 2": "95202885fae962f9e652783e111276bb"})
+        self.__checkDrawing(a3, 200)
 
     def testActionWithTwoDependents(self):
         a1 = self.__createMockedAction("a1", "a1", [], 12.5, 13.5, ActionTree.Action.Successful)
@@ -112,4 +112,4 @@ class Report(unittest.TestCase):
         a3 = self.__createMockedAction("a3", "a3", [a1], 14.0, 15.5, ActionTree.Action.Successful)
         a4 = self.__createMockedAction("a4", "a4", [a2, a3], 17.0, 17.5, ActionTree.Action.Successful)
 
-        self.__checkDrawing(a4, 200, {"Python 2.7, Cygwin - 1": "4c4362ef3df4b0f631f03afd03ef45e0", "Python 2.7, Cygwin - 2": "240a21cfd5fdd8c22db4640938cb3e8d"})
+        self.__checkDrawing(a4, 200)
